+++ Projects/App/WatchExtension/Sources/View/ControlsView.swift
@@ -1,45 +1,45 @@
+////
+////  File.swift
+////  AppWatchExtension
+////
+////  Created by 박현우 on 2023/06/27.
+////
 //
-//  File.swift
-//  AppWatchExtension
+//import SwiftUI
 //
-//  Created by 박현우 on 2023/06/27.
+//struct ControlsView: View {
+//    @EnvironmentObject var workoutManager: WorkoutManager
 //
-
-import SwiftUI
-
-struct ControlsView: View {
-    @EnvironmentObject var workoutManager: WorkoutManager
-
-    var body: some View {
-        HStack {
-            VStack {
-                Button {
-                    workoutManager.endWorkout()
-                } label: {
-                    Image(systemName: "xmark")
-                }
-                .tint(.red)
-                .font(.title2)
-                
-                Text("End")
-            }
-            
-            VStack {
-                Button {
-                    workoutManager.togglePause()
-                } label: {
-                    Image(systemName: workoutManager.running ? "pause" : "play")
-                }
-                .tint(.yellow)
-                .font(.title2)
-                Text(workoutManager.running ? "Pause" : "Resume")
-            }
-        }
-    }
-}
-
-struct ControlsView_Previews: PreviewProvider {
-    static var previews: some View {
-        ControlsView().environmentObject(WorkoutManager())
-    }
-}
+//    var body: some View {
+//        HStack {
+//            VStack {
+//                Button {
+//                    workoutManager.endWorkout()
+//                } label: {
+//                    Image(systemName: "xmark")
+//                }
+//                .tint(.red)
+//                .font(.title2)
+//                
+//                Text("End")
+//            }
+//            
+//            VStack {
+//                Button {
+//                    workoutManager.togglePause()
+//                } label: {
+//                    Image(systemName: workoutManager.running ? "pause" : "play")
+//                }
+//                .tint(.yellow)
+//                .font(.title2)
+//                Text(workoutManager.running ? "Pause" : "Resume")
+//            }
+//        }
+//    }
+//}
+//
+//struct ControlsView_Previews: PreviewProvider {
+//    static var previews: some View {
+//        ControlsView().environmentObject(WorkoutManager())
+//    }
+//}
