+++ Projects/App/WatchExtension/Sources/View/SessionPagingView.swift
@@ -1,49 +1,49 @@
+////
+////  SessionPagingView.swift
+////  AppWatchExtension
+////
+////  Created by 박현우 on 2023/06/27.
+////
 //
-//  SessionPagingView.swift
-//  AppWatchExtension
+//import SwiftUI
+//import WatchKit
 //
-//  Created by 박현우 on 2023/06/27.
+//struct SessionPagingView: View {
+//    @EnvironmentObject var workoutManager: WorkoutManager
+//    @Environment(\.isLuminanceReduced) var isLuminanceReduced
+//    @State private var selection: Tab = .metrics
 //
-
-import SwiftUI
-import WatchKit
-
-struct SessionPagingView: View {
-    @EnvironmentObject var workoutManager: WorkoutManager
-    @Environment(\.isLuminanceReduced) var isLuminanceReduced
-    @State private var selection: Tab = .metrics
-
-    enum Tab {
-        case controls, metrics
-    }
-
-    var body: some View {
-        TabView(selection: $selection) {
-            ControlsView()
-                .tag(Tab.controls)
-            
-            MetricsView()
-                .tag(Tab.metrics)
-        }
-        .navigationBarBackButtonHidden(true)
-        .onChange(of: workoutManager.running) { _ in
-            displayMetricsView()
-        }
-        .tabViewStyle(PageTabViewStyle(indexDisplayMode: isLuminanceReduced ? .never : .automatic))
-        .onChange(of: isLuminanceReduced) { _ in
-            displayMetricsView()
-        }
-    }
-
-    private func displayMetricsView() {
-        withAnimation {
-            selection = .metrics
-        }
-    }
-}
-
-struct PagingView_Previews: PreviewProvider {
-    static var previews: some View {
-        SessionPagingView().environmentObject(WorkoutManager())
-    }
-}
+//    enum Tab {
+//        case controls, metrics
+//    }
+//
+//    var body: some View {
+//        TabView(selection: $selection) {
+//            ControlsView()
+//                .tag(Tab.controls)
+//            
+//            MetricsView()
+//                .tag(Tab.metrics)
+//        }
+//        .navigationBarBackButtonHidden(true)
+//        .onChange(of: workoutManager.running) { _ in
+//            displayMetricsView()
+//        }
+//        .tabViewStyle(PageTabViewStyle(indexDisplayMode: isLuminanceReduced ? .never : .automatic))
+//        .onChange(of: isLuminanceReduced) { _ in
+//            displayMetricsView()
+//        }
+//    }
+//
+//    private func displayMetricsView() {
+//        withAnimation {
+//            selection = .metrics
+//        }
+//    }
+//}
+//
+//struct PagingView_Previews: PreviewProvider {
+//    static var previews: some View {
+//        SessionPagingView().environmentObject(WorkoutManager())
+//    }
+//}
